+++ R/conf.R
@@ -6,7 +6,9 @@
 #' @param from [[character]]
 #' @param dir_from [[character]]
 #' @param sep [[character]]
+#' @param sep_opt_name [[character]]
 #' @param inheritance_handling [[logical]]
+#' @param force_from_file [[logical]]
 #'
 #' @return [[list] or [character]]
 #'
@@ -16,14 +18,15 @@ conf_get <- function(
   from = "config.yml",
   dir_from = here::here(),
   sep = "/",
-  inheritance_handling = TRUE
+  sep_opt_name = "_",
+  inheritance_handling = TRUE,
+  force_from_file = FALSE
 ) {
-  configs <- getOption(from)
-  if (is.null(configs)) {
-    # configs <- config::get(file = here::here(dir_from, from))
-    # Per se okay but hard to test and inflexible for non-standard use cases
-
-    # Better:
+  # configs <- getOption(from)
+  pkg_this <- devtools::as.package(dir_from)$package
+  from_opts <- stringr::str_glue("{pkg_this}{sep_opt_name}{from}")
+  configs <- getOption(from_opts)
+  if (is.null(configs) || force_from_file) {
     configs <- config::get(file = fs::path(dir_from, from))
   }
 
@@ -60,21 +63,37 @@ conf_load <- function(
   pattern_disregard = "^(_|\\.|codecov|travis)"
 ) {
   config_files <- fs::dir_ls(dir, type = "file", regexp = "\\.yml$")
+
+  # Filter out special configs such as for TravisCI and `{covr}`:
   idx <- config_files %>%
     fs::path_file() %>%
     stringr::str_detect(pattern_disregard, negate = TRUE)
   config_files <- config_files[idx]
 
+  # Read config file content and put it into an actual option:
   config_files %>%
-    purrr::walk(function(.file) {
-      message(stringr::str_glue("Loading configs into options: {fs::path_file(.file)}"))
-      arg_list <- list(config::get(file = .file)) %>%
-        purrr::set_names(fs::path_file(.file))
-      rlang::call2(quote(options), !!!arg_list) %>%
-        rlang::eval_tidy()
-    })
+    purrr::walk(conf_load__inner)
 
   config_files
+}
+
+conf_load__inner <- function(
+  .file,
+  .sep_opt_name = "_",
+  .verbose = TRUE
+) {
+  if (.verbose) {
+    message(stringr::str_glue(
+      "Loading configs into options: {fs::path_file(.file)}"))
+  }
+  pkg <- devtools::as.package(".")$package
+  # opt_name <- fs::path_file(.file)
+  opt_name <- stringr::str_glue("{pkg}{.sep_opt_name}{fs::path_file(.file)}")
+  # print(opt_name)
+  arg_list <- list(config::get(file = .file)) %>%
+    purrr::set_names(opt_name)
+  rlang::call2(quote(options), !!!arg_list) %>%
+    rlang::eval_tidy()
 }
 
 #' Assign configs to local variables \lifecycle{experimental}
@@ -138,4 +157,48 @@ conf_assign <- function(
 conf_merge <- function(config_x, config_y) {
   # config::merge(config_x, config_y)
   conf_merge_lists(config_x, config_y)
+}
+
+# For `onAttach()` --------------------------------------------------------
+
+#' Load internal package configs \lifecycle{experimental}
+#'
+#' Put this in your custom `.onAttach()` function to load your package's
+#' internal configs based on `.yml` files that typically should live in your
+#' `./inst` directory (which becomes the root package directory once the package
+#' is attached).
+#'
+#' TODO-20191115-2: Explain env var `CONFX_AUTO_LOAD_INTERNAL`
+#'
+#' @param pkgname [[character()]] See [?.onAttach]
+#'
+#' @return [[list]]
+#' @export
+conf_auto_load_internal <- function(pkgname) {
+  env_auto_load_internal <- as.logical(
+    Sys.getenv("CONFX_AUTO_LOAD_INTERNAL", FALSE))
+  if (env_auto_load_internal) {
+    conf_load(system.file(package = pkgname))
+  } else {
+    list()
+  }
+}
+
+#' Load project configs \lifecycle{experimental}
+#'
+#' Put this in your custom `.onAttach()` function to load your project's configs
+#' based on `.yml` files that typically should live in your project's root
+#' directory (check via [here::here()]).
+#'
+#' TODO-20191115-3: Explain env var `CONFX_AUTO_LOAD`
+#'
+#' @return [[list]]
+#' @export
+conf_auto_load <- function() {
+  env_auto_load <- as.logical(Sys.getenv("CONFX_AUTO_LOAD", FALSE))
+  if (env_auto_load) {
+    conf_load()
+  } else {
+    list()
+  }
 }