+++ tests/testthat/setup-prerequisites.R
@@ -1,0 +1,2 @@
+PKG_THIS <- devtools::as.package(".")$package
+SEP_OPT_NAME <- "_"