+++ tests/testthat/test-conf_get.R
@@ -1,25 +1,35 @@
+dir_from <- test_path()
+
 test_that("conf_get() works (branch)", {
   # skip_on_travis()
-  res <- confx::conf_get("col_names", dir_from = test_path())
+  res <- confx::conf_get("col_names", dir_from = dir_from)
   expect_is(res, "list")
   expect_true(length(res) > 0)
 })
 
 test_that("conf_get() works (leaf)", {
   # skip_on_travis()
-  res <- confx::conf_get("col_names/.col_id", dir_from = test_path())
+  res <- confx::conf_get("col_names/.col_id", dir_from = dir_from)
   expect_is(res, "character")
   expect_true(length(res) == 1)
 })
 
 test_that("conf_get() works (all)", {
   # skip_on_travis()
-  res <- conf_get(dir_from = test_path())
+  res <- conf_get(dir_from = dir_from)
   expect_is(res, "list")
   expect_true(length(res) >= 1)
 })
 
 test_that("conf_get() works (error)", {
   # skip_on_travis()
-  expect_error(conf_get("notthere", dir_from = test_path()), "No such element in list: notthere")
+  expect_error(conf_get("notthere", dir_from = dir_from), "No such element in list: notthere")
+})
+
+test_that("conf_get() works (force from file)", {
+  # skip_on_travis()
+  warning("TODO-2019115-1: `force_from_file` hard to test, think about including an automatic load count or something similar")
+  res <- conf_get(dir_from = dir_from, force_from_file = TRUE)
+  expect_is(res, "list")
+  expect_true(length(res) >= 1)
 })