+++ tests/testthat/test-conf_load.R
@@ -15,10 +15,20 @@ test_that("conf_load() works (call)", {
 
 test_that("conf_load() works (after)", {
   # skip_on_travis()
-  res <- getOption("config.yml")
+  res <- getOption("confx_config.yml")
   expect_true(!is.null(res))
 })
 
-# Clean up -----
-options("config.yml" = NULL)
-options("config_2.yml" = NULL)
+test_that("Options are cleaned up", {
+  opt_name <- stringr::str_glue("{PKG_THIS}{SEP_OPT_NAME}config.yml")
+  arg_list <- list(NULL) %>% purrr::set_names(opt_name)
+  rlang::call2(quote(options), !!!arg_list) %>%
+    rlang::eval_tidy()
+  expect_identical(options(opt_name) %>% unlist(), NULL)
+
+  opt_name <- stringr::str_glue("{PKG_THIS}{SEP_OPT_NAME}config_2.yml")
+  arg_list <- list(NULL) %>% purrr::set_names(opt_name)
+  rlang::call2(quote(options), !!!arg_list) %>%
+    rlang::eval_tidy()
+  expect_identical(options(opt_name) %>% unlist(), NULL)
+})